+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning].
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-04-15
+
 ### Fixed
 
 - Fix a broken link in the API docs.
@@ -30,6 +32,7 @@ Initial release.
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/apyrgio/helvetia/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/apyrgio/helvetia/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/apyrgio/helvetia/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/apyrgio/helvetia/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/apyrgio/helvetia/releases/tag/v0.1.0